Add support for different attacks
New Attack Menu (not fully implemented) for selecting an attack after targeting an enemy, a new Attack class to hold attack properties like cost, damage, which die to roll, and damage. Players start with only one attack, with range 1, for now.

Added support for rolling dice anytime, but the roll only counts if it's the correct die for the chosen attack.